图片上传

+++ pinyougou-sellergoods-service/src/main/java/com/bees360/sellergoods/service/impl/ItemCatServiceImpl.java
@@ -74,7 +74,14 @@ public class ItemCatServiceImpl implements ItemCatService {
 	@Override
 	public void delete(Long[] ids) {
 		for(Long id:ids){
-			itemCatMapper.deleteByPrimaryKey(id);
+			TbItemCatExample example=new TbItemCatExample();
+			Criteria criteria = example.createCriteria();
+			criteria.andParentIdEqualTo(id);
+			List<TbItemCat> list = itemCatMapper.selectByExample(example);
+			System.out.println("size:"+list.size());
+			if(list.size()==0) {
+				itemCatMapper.deleteByPrimaryKey(id);
+			}
 		}		
 	}
 	
@@ -96,5 +103,14 @@ public class ItemCatServiceImpl implements ItemCatService {
 		Page<TbItemCat> page= (Page<TbItemCat>)itemCatMapper.selectByExample(example);		
 		return new PageResult(page.getTotal(), page.getResult());
 	}
-	
+		
+		//		根据parentID返回列表
+		public List<TbItemCat> findByParentId(Long parentId){
+			TbItemCatExample example=new TbItemCatExample();
+			Criteria criteria = example.createCriteria();
+			criteria.andParentIdEqualTo(parentId);
+			return itemCatMapper.selectByExample(example);
+			
+		}
+		
 }
